hilumxxxx

+++ test1kma/bai3.cpp
@@ -1,9 +1,10 @@
-/*Bài 3: Viết chương trình đọc và 2 số nguyên và in ra kết quả của phép (+), phép 
+/*Bài 3: Viết chương trình đọc và 2 số nguyên và in ra kết quả của phép (+), phép
 trừ (-), phép nhân (*), phép chia (/). Nhận xét kết quả chia 2 số nguyên. */
 #include <iostream>
 using namespace std;
 
-int main() {
+int main()
+{
     int a, b;
     cin >> a >> b;
 
@@ -12,7 +13,7 @@ int main() {
     cout << "a * b = " << a * b << endl;
 
     if (b != 0)
-        cout << "a / b = " << a / b << endl;
+        cout << "a / b = " << (float)a / b << endl;
     else
         cout << "Khong the chia cho 0\n";
